cleanup: remove redundant loads of various DKIM modules
These are cases where a high-level module loaded a dependent module
(such as Algorithm/rsa_sha1.pm) even though it already loads another
module which brings it in.

In another case, Signer.pm was loading a module it did not directly
use, but its base class used. So this patch makes the base class
load the module instead.